packages/py-moose-lib/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
docutils==0.21.2
idna==3.7
importlib_metadata==8.0.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
kafka-python-ng==2.2.2
keyring==25.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.3.0
nh3==0.2.17
pkginfo==1.10.0
Pygments==2.18.0
readme_renderer==43.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
setuptools==70.1.1
twine==5.1.1
urllib3==2.2.2
wheel==0.43.0
zipp==3.19.2
